Bring allow all feature to non-debug variant
We have completely dropped the redundant modules feature, so "Umount modules" will be useless, but no longer.

Allow all feature now uses global default value for "Umount modules" in App Profiles as a switch. Set it to OFF to ENABLE this feature.

If this feature is enabled, the Custom App Profile switch will be su BLACK LIST. Otherwise, the allow list is used as normal.

We will update the manager translations to make this feature more clarify later.